Moved size_t, ssize_t and int32_t to int64_t.
Instead of doing conversions from between size_t, ssize_t, int32_t and
int64_t, all support was moved to int64_t, with the exception of ssize_t
in the custom read function in VolumeArchiveLibarchive (libarchive
requirement).

Moved iterator to const_iterator.

BUG=chromium:409479

+++ unpacker-test/cpp/fake_lib_archive.h
@@ -37,7 +37,7 @@ extern const char* archive_data;
 
 // The size of archive_data.
 // By default it is set to 0, which forces failure for archive_read_data.
-extern size_t archive_data_size;
+extern int64_t archive_data_size;
 
 // Bool variables used to force failure responses for libarchive API.
 // By default all should be set to false.
